chatbot/backend/requirements.txt
# Bank Chatbot Backend Requirements

# Web Framework
fastapi==0.128.0
uvicorn[standard]==0.40.0

# Vector Database
chromadb==1.1.0
duckdb==0.9.2

# Embeddings
sentence-transformers==5.1.0
torch==2.10.0
torchvision==0.25.0
torchaudio==2.10.0

# LLM & NLP
requests==2.32.5
pyyaml==6.0.3

# Multi-Agent Framework
crewai==1.9.3
langchain==1.2.10
langchain-community==0.4.1
langchain-ollama==1.0.1
ollama==0.6.1
# openai==1.83.0
#apscheduler>=3.10.0
# email-validator>=2.0.0
# pydantic[email]>=2.0.0
#fastapi-sso>=0.9.0